+++ src-tauri/src/handler/calendar.rs
@@ -18,8 +18,8 @@ pub fn update_event(
     date: String,
     title: String,
     description: Option<String>,
-    start_time: i64,
-    duration: i64,
+    start_time: f32,
+    duration: f32,
     color: Option<String>,
 ) -> DbResult<()> {
     update_planner_event_in_db(
